feat: add validation for users, tracks

+++ src/track/track.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateTrackDto } from './dto/create-track.dto';
 import { UpdateTrackDto } from './dto/update-track.dto';
 import { ITrackData } from 'src/_db/Tracks';
 import dbManager from 'src/_db/dbManager';
+import { isUUID } from '../utils/isUUID';
 
 @Injectable()
 export class TrackService {
@@ -21,20 +22,38 @@ export class TrackService {
   }
 
   findOne(id: string) {
+    if (!isUUID(id)) {
+      throw new HttpException('Wrong id', HttpStatus.BAD_REQUEST)
+    }
+    if (!dbManager.trackExists(id)) {
+      throw new HttpException('Unknown record', HttpStatus.NOT_FOUND)
+    }
     return dbManager.getTrackById(id);
   }
 
   update(id: string, updateTrackDto: UpdateTrackDto) {
+    if (!isUUID(id)) {
+      throw new HttpException('Wrong id', HttpStatus.BAD_REQUEST)
+    }
+    if (!dbManager.trackExists(id)) {
+      throw new HttpException('Unknown record', HttpStatus.NOT_FOUND)
+    }
     let trackData: ITrackData = {
       name: updateTrackDto.name ?? null,
       artistId: updateTrackDto.artistId ?? null,
       albumId: updateTrackDto.albumId ?? null,
       duration: updateTrackDto.duration ?? null,
     };
-    return dbManager.updateTrack(id, updateTrackDto)
+    return dbManager.updateTrack(id, trackData)
   }
 
   remove(id: string) {
+    if (!isUUID(id)) {
+      throw new HttpException('Wrong id', HttpStatus.BAD_REQUEST)
+    }
+    if (!dbManager.trackExists(id)) {
+      throw new HttpException('Unknown record', HttpStatus.NOT_FOUND)
+    }
     return dbManager.deleteTrack(id);
   }
 }
