+++ src/user/user.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import dbManager from 'src/_db/dbManager';
 import { IUserData } from 'src/_db/Users';
+import { isUUID } from '../utils/isUUID';
 
 @Injectable()
 export class UserService {
@@ -19,10 +20,25 @@ export class UserService {
   }
 
   findOne(id: string) {
+    if (!isUUID(id)) {
+      throw new HttpException('Wrong id', HttpStatus.BAD_REQUEST)
+    }
+    if (!dbManager.userExists(id)) {
+      throw new HttpException('Unknown record', HttpStatus.NOT_FOUND)
+    }
     return dbManager.getUserById(id);
   }
 
   update(id: string, updateUserDto: UpdateUserDto) {
+    if (!isUUID(id)) {
+      throw new HttpException('Wrong id', HttpStatus.BAD_REQUEST)
+    }
+    if (!dbManager.userExists(id)) {
+      throw new HttpException('Unknown record', HttpStatus.NOT_FOUND)
+    }
+    if (!dbManager.checkPassword(id, updateUserDto.oldPassword)) {
+      throw new HttpException('Unknown record', HttpStatus.FORBIDDEN)
+    }
     let userData: Partial<IUserData> = {
       password: updateUserDto.oldPassword,
       newPassword: updateUserDto.newPassword
@@ -31,6 +47,12 @@ export class UserService {
   }
 
   remove(id: string) {
+    if (!isUUID(id)) {
+      throw new HttpException('Wrong id', HttpStatus.BAD_REQUEST)
+    }
+    if (!dbManager.userExists(id)) {
+      throw new HttpException('Unknown record', HttpStatus.NOT_FOUND)
+    }
     return dbManager.deleteUser(id);
   }
 }
